Potential fix for code scanning alert no. 30: Multiplication result converted to larger type
To fix the problem, we need to ensure that the multiplication is performed using a larger integer type to avoid overflow. We can achieve this by casting one of the operands to size_t before performing the multiplication. This will promote the entire expression to size_t, ensuring that the multiplication is done using the larger type.

The best way to fix the problem is to cast segment_width to size_t before the multiplication. This change should be made on line 287.

+++ libtiff-v4.0.10/libtiff/tif_webp.c
@@ -284,7 +284,7 @@ TWebPPreDecode(TIFF* tif, uint16 s)
   sp->sDecBuffer.width = segment_width;
   sp->sDecBuffer.height = segment_height;
   sp->sDecBuffer.u.RGBA.stride = segment_width * sp->nSamples;
-  sp->sDecBuffer.u.RGBA.size = segment_width * sp->nSamples * segment_height;
+  sp->sDecBuffer.u.RGBA.size = (size_t)segment_width * sp->nSamples * segment_height;
   
   if (sp->nSamples > 3) {
     sp->sDecBuffer.colorspace = MODE_RGBA;
